Refactor usePinDialog hook;
